Add SandboxItem component for flexible content display.
